Add clang-tidy support and fix all lint warnings
Add two clang-tidy profiles: .clang-tidy (dev/IDE, advisory) and
.clang-tidy-lint (CI, warnings-as-errors). Wire clang-tidy into the
build via a Tidy CMake config that runs on verify_interface_header_sets,
and add a `make clang-tidy` target integrated into `make lint`.

Fix all clang-tidy findings across headers and tests:
- Use in-class member initializers instead of constructor init lists
  where the value is always the same default (modernize)
- Move deleted copy operations to public section for clearer diagnostics
- Split comma-separated variable declarations
- Add missing parentheses for operator precedence clarity
- Replace `auto*` with `const auto*` where appropriate
- Use `std::max` instead of manual comparison
- Introduce `iconv_error_rc` constant for iconv error checks to avoid
  signed/unsigned comparison warnings — std::cmp_equal is semantically
  wrong here since iconv returns SIZE_MAX, not mathematical -1
- Add missing `<utility>` and `<algorithm>` includes

+++ tests/beman/transcode/iconv_real.test.cpp
@@ -9,6 +9,7 @@
 #include <beman/transcode/iconv_transcode_or_error_view.hpp>
 
 #include <array>
+#include <cstddef>
 #include <span>
 #include <string>
 #include <vector>
@@ -87,10 +88,10 @@ TEST_CASE("real iconv E2BIG: long input with small buffer", "[transcoding::iconv
         output.push_back(c);
     REQUIRE(output.size() == 400);
     for (int i = 0; i < 100; ++i) {
-        CHECK(static_cast<unsigned char>(output[i * 4]) == 0x41);
-        CHECK(static_cast<unsigned char>(output[i * 4 + 1]) == 0x00);
-        CHECK(static_cast<unsigned char>(output[i * 4 + 2]) == 0x00);
-        CHECK(static_cast<unsigned char>(output[i * 4 + 3]) == 0x00);
+        CHECK(static_cast<unsigned char>(output[static_cast<std::size_t>(i) * 4]) == 0x41);
+        CHECK(static_cast<unsigned char>(output[(static_cast<std::size_t>(i) * 4) + 1]) == 0x00);
+        CHECK(static_cast<unsigned char>(output[(static_cast<std::size_t>(i) * 4) + 2]) == 0x00);
+        CHECK(static_cast<unsigned char>(output[(static_cast<std::size_t>(i) * 4) + 3]) == 0x00);
     }
 }
